+++ invenio_notify/services/__init__.py
@@ -1,0 +1,36 @@
+from .config import (
+    NotifyInboxServiceConfig,
+    EndorsementServiceConfig,
+    ReviewerMapServiceConfig,
+    ReviewerServiceConfig,
+    EndorsementRequestServiceConfig,
+    EndorsementReplyServiceConfig,
+)
+from .service import (
+    BasicDbService,
+    NotifyInboxService,
+    EndorsementService,
+    ReviewerService,
+    ReviewerMapService,
+    EndorsementRequestService,
+    EndorsementReplyService,
+)
+
+__all__ = [
+    # Services
+    "BasicDbService",
+    "NotifyInboxService",
+    "EndorsementService",
+    "ReviewerService",
+    "ReviewerMapService",
+    "EndorsementRequestService",
+    "EndorsementReplyService",
+
+    # Configs
+    "NotifyInboxServiceConfig",
+    "EndorsementServiceConfig",
+    "ReviewerMapServiceConfig",
+    "ReviewerServiceConfig",
+    "EndorsementRequestServiceConfig",
+    "EndorsementReplyServiceConfig",
+]
