+++ invenio_notify/services/service/__init__.py
@@ -1,0 +1,20 @@
+from .base_service import BasicDbService
+from .endorsement_reply_service import EndorsementReplyService
+from .endorsement_request_service import EndorsementRequestService
+from .endorsement_service import EndorsementService
+from .inbox_service import NotifyInboxService
+from .reviewer_map_service import ReviewerMapService
+from .reviewer_service import ReviewerService
+
+__all__ = [
+    # Base classes and utilities
+    "BasicDbService",
+
+    # Services
+    "NotifyInboxService",
+    "EndorsementService",
+    "ReviewerService",
+    "ReviewerMapService", 
+    "EndorsementRequestService",
+    "EndorsementReplyService",
+]
