+++ invenio_notify/services/service/endorsement_request_service.py
@@ -1,0 +1,5 @@
+from .base_service import BasicDbService
+
+
+class EndorsementRequestService(BasicDbService):
+    """ Endorsement Request Service, no use for now """
